requirements.txt
Flask[async]>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv
flask-socketio>=5.1.1
Flask-Dance>=5.0.0
selenium>=4.1.0
webdriver-manager>=3.8.0
pandas>=1.3.0
plotly>=5.3.0
APScheduler>=3.8.1
python-telegram-bot>=20.0
aiohttp>=3.8.1